+++ backend/app/config/db.config.js
@@ -1,0 +1,8 @@
+const mysql = require("mysql2");
+
+module.exports = mysql.createConnection({
+    host: 'localhost',
+    user: 'root',
+    password: 'pz3YIiTtX1%:K5J',
+    database: 'groupomania',
+});
