+++ groupomania/src/components/ArticlePost.js
@@ -5,7 +5,8 @@ import postRequest from "../services/post-request";
 
 function ArticlePost ({setArticleList}) {
     const [newArticle, setNewArticle] = useState({title: '', content: '', author: ''})
-
+    
+    //voir autre méthode pour mettre à jour la liste d'article
     const submitArticle = (newArticle) => {
       if (newArticle.title==='' || newArticle.content==='' || newArticle.author==='') {
         alert("votre article est vide ! Veuillez remplir tous les champs !")
